Configure test's fileStoreService to use relative path
The old fileStoreService bean config uses
    "${project.basedir}/target/test-classes/filestore"
as the path to simulate the file store on the local file system. The new,
Java-based config omits the property uses the path
    "target/test-classes/filestore"
relative to the execution environment.

Potential problems include the possibility that this will not work the
same way in other contexts, and that the configuration's bean-getter has
the minor side effect of creating the directory if needed.