Debug author-name parsing with XPath
XPath queries that are local to a node now correctly search only within
the node, rather than exploring the node's context (i.e., parent and
siblings) and finding stuff outside of the scope that we want.

XmlToArticle now correctly parses names from the lists of authors and
editors. Test coverage has been added to ensure that non-empty names are
found and don't repeat each other (as they did before the bug was fixed).