Major refactor of article-XML-reading code
The old structure of XmlToArticle, however lovely and lambda-like, was
getting unwieldy. The verbosity (and maybe the eventual performance
overhead) were outweighing the conceptual benefits of representing each
field-setter with an anonymous class.

The replacement is the ArticleXml class and friends. It instead represents
one XML file with one object, eschews the earlier object-per-field
approach, and does everything much more procedurally. The XPath
expressions are gathered in one method (setFromXml), which dispatches
to parsing methods where needed. Much cleaner this way.