Debug Spring's extension-stripping on DOIs
It is no longer necessary to append a dummy extension onto the end of a
DOI to get proper behavior (e.g., "journal.pone.0000.doi" -- this was
not reflected in any formal test code but was habitual for manual testing
with the Python client script).

The Python test client has been fleshed out a bit and provided with an
example invocation, now that we can do so without the workaround testing
for wrong behavior.