Partial implementation of article asset creation on ingest
This development line assumes that we want ArticleAsset entities to be
created at the same time that an article is ingested, and that the payload
data for those assets will be provided later for the file store.

The implementation currently finds figures and tables, but there are bugs
(commented) in the XPath to find supplementary material references and
inline graphics.