Change ArticleCrudService.read to return an InputStream
It is more consistent with the create and update methods to pass streams
back and forth with the controller. (Thanks to John Callaway for pointing
this out.)

The controller still dumps the stream into a byte array before returning it
in a ResponseEntity. Using an HttpServletResponse stream seems to require
doing the same. Should look later into whether the streams can be piped for
more efficiency.