Partially implement asynchronous asset-uploading with delayed write
This is the simplest implementation for attaching assets to articles, which
delays creating database entities for the assets until the actual data is
uploaded, and rereads the article as necessary to get asset metadata from
it. This rereading is very inefficient and needs attention. So far, the
implementation writes the assets only to the database, not the file store,
and does not fill in the extension column (the file type). Only 'create' is
supported out of the CRUD operations.

There was also a lot of incidental refactoring around the service and
controller.

The new code needs unit test coverage, but the Python scripts for manual
testing have been updated to exercise it.