Change API to use extensions on object IDs, and refactor to model IDs
Under the API change, entity URLs now have a file extension appropriate to
their file store data. This clarifies the relation between assets and
what's in the file store, and will make READ requests behave more sensibly
when they come from a web browser or wget.

The goal of the refactoring was to avoid using strings as DOIs/IDs wherever
the expected DOI format was unclear. The new ArticleSpaceId class contains
the DOI and yields it in a particular format (i.e., prefixed with
"info:doi/" or not) close to where that format is required. It also bundles
the file extension and asset parent with the DOI. There was extensive
refactoring in the controller, services, and tests to use this class
instead of bare strings.