Allow more characters in file name extension
The code in ArticleSpaceId that parses identifier elements from a URL is
now more permissive about non-word characters (such as hyphens and
underscores) in file name extensions. Some incidental refactoring was done
in the same class. The client now sees articles and assets identified with
the file extensions (the same way they would type in a URL).