Refactor Python test script to store its own case data
Maintaining a programmatic list of test cases will be preferable to passing
the cases from the command line in the near future, in order to test in
different logical domains (for example, articles versus issues). Otherwise
it would be necessary to have different scripts/interfaces for different
test types, even though we generally want to run them all at once.