Change volume creation from PUT to POST, add to test script (needs work)
Spring seems not to like POSTs with no associated data -- will need to look
into what is proper RESTful design in this context. The test script now
should fully exercise volume and issue creation, but does not currently
work. (The server behaves as expected, but the client script hangs after
the response has been passed to it.)