Add message body to Request class, with default to empty
This allows the REST client library to have a message body defined, which
was previously possible only through form fields.

When a READFUNCTION is not set at all for a pycurl object, it hangs
silently if the server tries to open the message body stream. This is a bad
failure mode and has already been an annoyance in development. Hence, the
client library defaults to streaming an empty string as the message body if
the server asks, even if the client user didn't define a body.