Unify ArticleCrud's create and update methods into upload, data in req body
This allows more proper use of the HTTP PUT method, such that it is always
valid whether an entity exists at the URL or not. The file is now uploaded
in the request body, which is properly symmetric with GET (it was uploaded
in a multipart file before, which was always a bad kludge).

Analogous refactoring of the other create and update methods will follow.