Replace AssetCrudService's create and update with upload
As with ArticleCrudController, the asset data is now uploaded in the
message body rather than in a multipart file. Newly irrelevant parts of the
superclasses are removed.

This is a partial implementation. When an asset is created, it does not
check that the article exists in the database (it only uses the article ID
to read the file store). Updating an asset without providing an "assetOf"
parameter is not supported yet. When an asset is updated but the "assetOf"
is provided anyway, it should check that it's consistent with the actual
parent article.