Add volume read method that provides metadata as JSON
More extensive API refactoring will be needed to expose article and asset
metadata the same way. Such requests will need to be disambiguous from GET
requests that just look at the raw data, and should provide a way to
specify the metadata format (in case we ever want to use something in
addition to JSON).