Some refactoring anticipating XML-as-asset modeling
We want to treat articles as entities separate from their XML data, such
that a request directed at an article refers to its metadata (the stuff in
the database) and its XML file is referred to as an asset. As preliminary
back-end refactoring, FileStoreController has been pulled down into
AssetCrudController (because only assets are in the file store now) and
ArticleCrudController has placeholders for its intended functionality.