Refactor DoiBasedIdentity type hierarchy; fix article/XML confusion bug
Previously, there was a bug with a few article operations, including
deletions, caused by a DoiBasedIdentity for an article's XML-holding asset
being confused with the identity that refers to the article's JSON
metadata. The root cause was that the same DoiBasedIdentity class
represented both, with runtime state indicating whether the object had an
associated object in the file store.

No more! The bug is fixed, and object-oriented salt has been sown in the
earth so that no such bug shall ever grow back. Asset identities now have
their own class, and methods that refer to file store data appear only
there. Appropriate refactoring has been done in the controller and service
layers.