Prefer empty lists of article XML data over null
Removed a lot of code that represented empty lists from article XML as
null, or checked the null cases. The null representations were originally
motivated by skittishness around Hibernate, but it should be fine. (And if
it doesn't, it ought to be refactored anyway, since using "null" to mean
"no elements" is a terrible antipattern.)