Reintroduce POST with multipart file argument, with "file" key
To get POST working for now, implementing as it was earlier, with the
article XML uploaded as a request parameter (whose type is multipart file)
rather than as the message body. It would be better to POST with the XML as
a simple message body if HTTP permits; will look into this.