Use requests lib in rhinotest.py; simplify
Tore a lot of code out of rhinotest.py and rebuilt simple create, read, and
delete cases for article. Will reimplement for assets, volumes, and issues.
The Python classes that represent test cases are in their own files now.

Deleted restclient.py entirely because the requests lib should obviate it.