Have ArticleXml look for NLM 3 citations ("element-citation")
Took out code to look for other two citation node types ("citation" and
"nlm-citation"); I don't know if they'll need to be supported, but I'll
assume not until it shows up in a test case.

Also, made PersonName null-tolerant again because the JSON test cases have
null suffix fields for CitedPersons but not ArticlePersons. This is in fact
a faithful representation of the old ingestion code. Will need to figure
out whether Rhino needs to reproduce this, or can safely introduce
consistency.