Be stricter about null vs "" in testing; test individual PersonName fields
Ambra uses null for the suffix of ArticlePerson and PersonName objects
inconsistently, so Rhino should reproduce and test for the behavior until
we're sure it's safe to do otherwise. For String fields checked in
IngestionTest, treat "" and null as unequal. The exceptional strkImgURI
case that was identified previously is now handled specially.

To make results clearer, PersonName's fields are now compared as individual
cases. Before, it used PersonName.equals, which didn't make null-vs-""
mismatches clear in the log output.