Have article POSTs respond with a copy of the received data, as JSON
This entailed refactoring the WriteResult type (and extracting it to its
own file) to contain the written object in addition to describing whether
the write was a create or update.

The current implementation involves reloading the Article object from
Hibernate, as the created object cannot initialize
Article.journals.volumes. If this can be fixed, it would save a trip to the
database on each article ingestion.