Partial implementation of asset metadata reading after uploading
Chaned the AssetCrudService interface to stream asset JSON to the HTTP
response, as with ArticleCrudService. AssetCrudController now sends the
JSON of the new asset in the response after an asset is posted.

The Python script currently exposes a bug with AssetMetadataController's
read method (seems to be controller-level) that will need fixing.