Add asset to represent XML file on article ingestion
IngestionTest now shows no failures with its initial case. Although
IngestionTest does not check file size, the new XML asset actually should
have its size set because the file is always available at ingestion, so
this is a TODO.

Also added a TODO about an assumption in ArticleCrudServiceImpl that each
asset in an existing article has a unique DOI. Actually it's the
DOI-extension pair that's unique. The assumption is true when ingesting a
new article but might cause errors if the same method is used to update an
existing article object.