Generalize whitespace handling on ingested text fields
"Collapse and trim whitespace" is now the default for all strings and
string lists read by the AbstractArticleXml subclasses. This fixes a
failing unit test with the CitedArticle.note field in the pone.0038869
case, and will prevent storing unnecessary whitespace in general.

Updated IngestionTest to permit such variations in whitespace when
comparing List<String> objects.