TNG-44: Efficiency hack for creating assets with direct SQL
The goal here was to eliminate the need to create and persist an entire
Article object, with its entire set of ArticleAsset objects, each time one
new ArticleAsset was to be created.

This fixes the target case (from the Python dummy client) but currently
breaks a few unit tests -- will fix.

Updated comments, and changed a method signature for readability.