Midway through refactoring to handle duplicate asset nodes
The pmed.0030297 case demonstrates that an asset can be repeated across
multiple XML nodes, and that we should create one asset to represent all of
them as long as the fields match. This is an incremental commit for that
refactoring task. It is touching on the findAllAssetNodes method and
presenting opportunities to use the DOI-to-node maps more efficiently.