Fix infinite recursion bug when writing to JSON in certain cases
Attempting to read pcbi.0020083 (once ingested) as JSON was causing Gson to
infinitely recur through Article.relatedArticles ->
ArticleRelationship.parentArticle -> Article. This field is now excluded in
the configuration of the entityGson bean (the output should still be
unambiguous).

Added pcbi.0020083 as a test case and added the JSON-writing operation to
the IngestionTest procedure, which should expose the bug that this commit
fixes.