TNG-54: Implement JSONP support; introduce ResponseReceiver class
The controller methods that read JSON metadata now will provide it in JSONP
format if the request provides a "callback" parameter.

The ResponseReceiver classes support JSONP and address the problem of
passing HttpServletResponse objects down into the service layer where they
don't belong. The response objects were passed this way to delay calling
getWriter() in order not to interfere with Spring printing error messages
in case of an exception. The ResponseReceiver provides an abstraction
around this, and the ServletJsonpResponse implementation supports JSONP
transparently.

Most of the modified files in this commit are changes to metadata-reading
method signatures. The interesting stuff is in
src/main/java/org/ambraproject/rhino/util/response/