TNG-85: Switch article.journals JSON view to a keyed list
This causes it to match the format for reading journals (rather than a flat
array). Currently, uses the non-associative view, which will keep lots of
volume and issue data out of the output. (To do: change the Hibernate
queries that are feeding the model into that view, so they aren't fetching
data that isn't used.)