Make issue article values more consistent with the rest of the interface
An issue object now has an "articleOrder" value, which is an array of the
identifiers of the articles in the issue. The "DoiList" object view cannot
be used for this purpose: although that view preserves the DOI order for
human-friendliness, the members of a JSON object are unordered by
definition. So it needs to be a plain array of DOIs.

We may still want to show actual article objects within the issue view.
There could be an "articles" object containing the actual article values
(in the normal key-value format) alongside the "articleOrder" array.

Also fleshed out the identity-vs.-key concepts within DoiBasedIdentity.