Set title and description when setting an issue's image article
This will lead to unwanted behavior if the image article is reingested: the
issue's title and description will be stale. The user-friendliest thing to
do would be to check for associated issues every time an article is
altered. Possibly a more practical option is to expose a feature to refresh
the issue.

The issue-refreshing action will need its own feature, since resubmitting
the same imageUri would cause no change. The only existing workaround would
be to change imageUri to a dummy value and then change it back.