Introduce a new view of assets for the article.assets field
Trying out a new way of listing assets as child properties of an article
with more focus on their individual significance. It's work in progress: it
still needs to be aligned with "assets" GET and "articles" POST/PUT
methods. It may eventually replace everything currently in
org.ambraproject.rhino.view.asset.raw.