Generalize writeJson method and pull format value into it
Allows service methods to be more agnostic about the eventual output
format. (It's still defined in the service layer, but fewer methods need to
know about it this way.)

Done in anticipation of a change from defining the metadata output format
in a URL parameter to defining it in a request header.