Remove HttpServletRequest args from controllers where possible
Introduced a new factory method in ServletResponseReceiver that takes the
JSONP callback parameter as an argument. Replaced HttpServletRequest
arguments in controller methods with direct access to the JSONP callback
and Accept header if that was all that the controller required. Left
HttpServletRequest arguments in place where they were needed to access the
full request URI, in order to parse out a DOI-like identifier.