Prevent exception-swallowing on mid-ingestion self-cleanup
Addresses the case where an ingestible archive fails validation while
adding asset files, and then a server-side exception is thrown while
deleting the already-persisted article data. With this change, the user
would still see a 500 error (as they should), but the original 400 error
will now make it to the logs.