AMEC-2216: Ensure contextElement is set on assets from XML
Adds logic to the code that creates thumbnail assets, to copy the
contextElement value from the uninitialized asset object. This means that
the value originally saved from article XML is preserved.

This obviates inferring the contextElement from the asset name (relying on
PLOS's naming convention), which was the cause of the documented bug with
correction articles. Removed code that does so, and the corresponding unit
test.