Fix HibernateOptimisticLockingFailureException on re-ingestion
In the ArticleCrudServiceImpl.modifyRelatedArticles workaround, modifying
the ArticleRelationship objects in place suffices to solve the "duplicate
rows" problem. The redundant hibernateTemplate.delete (when a relationship
is removed from XML between ingestions) actually causes an error.

Added a comment explaining that the code doesn't distinguish between
deleted and reciprocal relationships, but it's not a bug.