AMEC_2371: lazily load Article.citedArticles for performance boost.
In order to enable lazy loading, we have to use spring transactions (since otherwise, the hibernate session will be closed when we try to lazily load).  This led to a lot of changes and I ended up just adding transaction annotations on all controller methods.

Also refactored some of our hibernate config, in which things were specified twice (once in the XML and once in the annotations), leading to various bugs.

TODO: lots more manual testing
TODO: fix the unit tests so they run transactionally.  This was a world of pain.