By default, don't buffer JSON into memory before serializing
It was a more useful default to buffer it during development, but as long
as we don't expect exceptions to be thrown when building response JSON,
this is a better default.

The behavior can still be changed in individual Transceiver overrides.
Would be nice eventually to make it globally configurable at runtime.