DPRO-63: Remove dependence on FileStoreService.listFiles
This does approximately the same thing but leaves behind a few edge cases
where files that don't correspond to assets can be stored from the original
ingestible but not deleted if the article is disabled. At present, this
seems to be the best available compromise if FileStoreService does not
support listFiles, because only it (and nothing in the database) captures
the relationship between the article and those files.