DPRO-131: Fix/optimize query for current issue lastModified timestamp
The query's lastModified date needs to account for changes in the parent
volume (which is included in the returned view) in addition to the issue.
Query for both dates. This also is probably more efficient than loading the
full Issue object in the case where it returns a "Not Modified" response.