Simplify If-Modified-Since checks
The content repository interface provides timestamps for asset files, which
was not possible through FileStoreService. This obviates handling article
XML specially, and lets us support the If-Modified-Since header on all
assets, not just article XML.

We now query for crepo metadata on all asset requests, and no longer only
for reproxying cases.