DPRO-924: Remove ArticleCrudService.write; replace it in unit tests
ArticleCrudService.write was for creating a partial article by uploading
its XML manuscript without assets. We will not be supporting this under the
versioned model.

Replaced it in unit tests with a utility method that builds a small
ingestible archive around a single XML file. This will break unit tests
until write services have caught up with reads for the new data model.