DPRO-897: fix for articles not rendering in wombat.
pgen.0030208 includes XML in the authors section such as this:

<contrib xmlns:xlink="http://www.w3.org/1999/xlink" contrib-type="author" xlink:type="simple">
  <name name-style="western">
    <surname>Trinklein</surname>
    <given-names>Nathan D</given-names>
  </name>
  <xref ref-type="fn" rid="n103">
    <sup>
      <sup>¤</sup>
    </sup>
  </xref>
  <xref ref-type="aff" rid="aff1"/>
</contrib>

Notice the nested <sup> elements surrounding the ¤ character.  This
appears to be an aberration; most XML includes only the <sup>¤</sup>.
This leads to

net.sf.saxon.trans.XPathException: A sequence of more than one item is not allowed as the first argument of contains() ("", "")

The xpath expression that I've modified in this commit is identical in
the ambra code.  However, in that code, the call to the xpath is
enclosed in a large try-catch block, where the exception is logged.
This causes the article to render in ambra, but incorrectly, with only
the authors preceding the aberrant <sup> element displaying.  In
wombat, the exception is thrown instead of being swallowed, leading to
a 500 instead (before this commit).