DPRO-1079: Adapt services and controllers to use Archive
This will be necessary to support parallel ingestion into the versioned
data schema.

Removed the API endpoints that write an individual file (instead of
ingesting a whole archive). These have been softly deprecated for a while
and will become completely unsupportable once we are doing parallel
ingestions to the versioned data schema.