DPRO-1079: Fix for assets in manifest butnot manuscript
Resolves an error that happens if an ingestible contains an asset that is
only a "striking image" and is never mentioned in the manuscript. Its
absence from the manuscript prevented it from being assigned an asset type.

Refactored to clean up code around assigning asset types. As a side effect,
the `buildFromAssetMetadata` method no longer calls the same `getAssetType`
method, which introduces a TODO to read the asset type from the JSON
metadata instead. Will fix this when there is a unit test that exercises
the `buildFromAssetMetadata` method.