DPRO-1679: Present volume and issue views as lists instead of maps
The former API was obscuring the order of the issue list within each volume
and the volume list within each journal. The order is a persisted value and
we need to provide it one way or another.

This breaks backwards compatibility in principle, but there should be no
Wombat code that depends on this.