DPRO-1992: Initial sketch of migration script in Python
As a starting point, this replicates the behavior of Ambra's bootstrap
migrator. We might want to change the behavior to something more
sophisticated.

We might also want to copy Ambra's old migration scripts to here, and set
up the JSON file to point at them.

Needs documentation and better error-handling. Also, it is not inserting
timestamps correctly.