DPRO-2107: Initial sketch of a massive simplification of asset ingestion
The goal is to remove support for uploading asset files individually and to
get rid of the extra steps and kludges that it made necessary. Most
importantly, we avoid creating any ArticleAssets in an incomplete state and
delay persisting anything until the Article and all its assets are ready.

Ingestion basically runs without errors but has a lot of gaps and bugs. The
code does not correctly handle the manuscript asset, striking images, or
supp info. Some unit test code is also dummied out.