DPRO-2234: Improve exception handling
Reverted handling on individual terms with invalid syntax, so that we no
longer omit only that term and persist the rest.

Changed exception-swallowing in populateCategories so that errors cause the
"repopulateCategories" service to respond with a 500 status. But,
exceptions (even unanticipated RuntimeExceptions) from category population
will not halt an ingestion.

Rearranged custom exception types for clarity.