DPRO-2107: Create asset for stand-alone striking images when needed
Was able to cut out code that inspected asset DOIs for ".strk." because it
affected only the decision whether to create new ArticleAsset objects,
which is never necessary if the assets exist in the first place. The
article's setStrkImgURI has been set only from the manifest for a while, so
apparently we don't need the code there either.