DPRO-2107: Disable testArchiveError
This tested the exception-handling of a condition where an ingestible
archive contained a file not referenced by the manuscript or manifest. The
original exception was a side effect because the now-obsolete workflow
wanted to match each of those files to metadata from the manuscript and
couldn't proceed otherwise.

There would be little value in writing new code just to ensure that we
loudly fail in a case that no longer is significant or expeted. For now,
the service silently ignores extra files in the zip archive.