DPRO-2442: Groundwork to require manifest to list all archive files
Added the actual check and throw, though it will not be possible for
manifests to comply with the check until we expand the contract to allow
the manifest to list non-corpus files.

Deleted a kludge to keep some non-corpus files from causing errors, in
anticipation that they will be represented differently.