DPRO-382: Replace Transceiver with ServiceResponse
Solves a race condition where a destructive operation would be reported in
a response before it was actually persisted, and cleans up a lot of other
technical debt and needless code around Transceiver.

Caching functionality will be degraded until the deprecated calls to
CacheableServiceResponse.asJsonResponse(Gson) are replaced to pass the
"If-Modified-Since" timestamp from the controller layer to the
CacheableServiceResponse objects.