AMBR-218 Eliminate n+1 queries when fetching ingestion items
Previously, retrieving the items was subject to an n+1 query issue. For every
file a request would be made to fetch the item as well. This ensures that all
items are fetching at once with the files in one query.

This reduces the time it takes to do a request to /ingestions/{n}/items by about
1/3.