AMBR-1035 Simplify related link loading
Now that we canonicalize all links in the API to be "inbound", we
don't need to capture both sides of the links, which eliminates the
original need for this (see
https://jira.plos.org/jira/browse/DPRO-3006).

If we ingest an article with a relation to a not-yet-ingested article, it
will not persist the relation to the db (because the not-yet-ingested
article has no `articleId`). However, when the not-yet-ingested
article is ingested, we will persist the relation.

If we ingest an article with a relation to a not-yet-ingested article,
and that not-yet-ingested article does not include the reciprocal
link, that will be a problem, but it shouldn’t happen.